Move FormatUtils to core-domain, remove shared dep from core-ui
