+++ androidApp/src/main/kotlin/com/riox432/civitdeck/ui/detail/FileDownloadRow.kt
@@ -26,8 +26,8 @@ import com.riox432.civitdeck.R
 import com.riox432.civitdeck.domain.model.DownloadStatus
 import com.riox432.civitdeck.domain.model.ModelDownload
 import com.riox432.civitdeck.domain.model.ModelFile
+import com.riox432.civitdeck.domain.util.FormatUtils
 import com.riox432.civitdeck.ui.theme.Spacing
-import com.riox432.civitdeck.util.FormatUtils
 
 @Composable
 fun FileDownloadRow(
