+++ desktopApp/src/jvmMain/kotlin/com/riox432/civitdeck/ui/detail/DesktopDetailScreen.kt
@@ -73,7 +73,7 @@ import com.riox432.civitdeck.ui.theme.CornerRadius
 import com.riox432.civitdeck.ui.theme.Spacing
 import com.riox432.civitdeck.ui.theme.shimmer
 import com.riox432.civitdeck.ui.theme.Elevation
-import com.riox432.civitdeck.util.FormatUtils
+import com.riox432.civitdeck.domain.util.FormatUtils
 
 @Composable
 fun DesktopDetailScreen(
